+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+# 0.7.0
+
 - Fixes the link in the gist view will be opened in the iFrame.
 - Use the `postMessage` mechanism to set the iFrame's height, which should be much secure.